Use single class
